Add filter to function email_exists()
https://core.trac.wordpress.org/ticket/51379

As described in the ticket above, I'd like to have a filter in `email_exists()` in order to have more freedom when working with a third party system that is the leading system for my user-data. 

The filter is working the same way as the filter in `username_exists` in the lines above which was added in 4.9.0.